Enhanced the multiplication in Weyl algebra.
Added and modified some new functions related to plotting.
(sorry, not documented yet)

ifplot(), conplot(), plot() : push the 'index' which indicates
the index of the created canvas. A canvas is specified by (Id,Index),
where Id is the identifier for an ox server.

open_canvas(Id): create a white canvas.

draw_obj(Id,Index,PointOrSegment): create a white canvas. This function
also pushes the index.

PointOrSegment : a list. If the length is 2, it is regarded as a point.
If the length is 4, it is regarded as a line segment.

clear_canvas(Id,Index): clears the canvas.